+++ apps/web/src/services/alert-manager/AlertManagerLNB.vue
@@ -12,7 +12,7 @@ import { i18n } from '@/translations';
 
 import {
     filterLNBMenuByAccessPermission,
-} from '@/lib/access-control/page-access-permission-helper';
+} from '@/lib/access-control/page-access-helper';
 import { MENU_ID } from '@/lib/menu/config';
 import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
