+++ apps/web/src/services/asset-inventory/AssetInventoryLNB.vue
@@ -28,7 +28,7 @@ import { FAVORITE_TYPE } from '@/store/modules/favorite/type';
 
 import {
     filterLNBMenuByAccessPermission,
-} from '@/lib/access-control/page-access-permission-helper';
+} from '@/lib/access-control/page-access-helper';
 import { MENU_ID } from '@/lib/menu/config';
 import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
