+++ apps/web/src/services/my-page/MyPageLNB.vue
@@ -32,7 +32,7 @@ import { i18n } from '@/translations';
 
 import {
     filterLNBMenuByAccessPermission,
-} from '@/lib/access-control/page-access-permission-helper';
+} from '@/lib/access-control/page-access-helper';
 import { MENU_ID } from '@/lib/menu/config';
 import { MENU_INFO_MAP } from '@/lib/menu/menu-info';
